+++ lib/easy_caddy.rb
@@ -4,6 +4,7 @@ require 'pathname'
 require 'fileutils'
 
 require_relative 'easy_caddy/version'
+require_relative 'easy_caddy/error'
 require_relative 'easy_caddy/paths'
 require_relative 'easy_caddy/site'
 require_relative 'easy_caddy/registry'